Get current events at the time stream is opened
The semantics of current events stream is as follows: delivered events
are those that were present at the time the stream was opened.
Guarantee is that all events present during the entire stream life will
be delivered. If elements are removed before delivery, they might not be
delivered at all (due to batching).