Fix pagination for events by persistence id
Always query the event between the last known sequence number and the
upper limit, instead of the sequence number in a page range after the
last known number. Using the `limit` parameter to the query makes it
possible to retrieve batches of the desired size.

The previous method ignored higher number if the lower sequence numbers
were missing, and considered the source done so far.

Fix #11